requirements.txt
tensorflow==2.10.1
numpy
opencv-python
matplotlib
streamlit
scikit-learn
Pillow
scipy
